Keep events out of model
